Version 0.208
FIXES:

Fixed the tray menu item "Edit This Script".

Fixed compatibility-check for AutoIt2: some of the default AutoIt2 behaviors weren't being used.

Fixed WinWait commands: Exclude-Text and Exclude-Title weren't being handled correctly.

Turned off debug logging: If you were running an older version, you may have a small text file "C:\AutoHotkey SetForegroundWindowEx.txt", which you can safely delete after upgrading to this version.

Changed StringReplace so that it doesn't trim whitespace from the result, which matches AutoIt2's behavior.

CHANGES:

Included the VC++ project in the source code ZIP file.

Changed the default for .aut (AutoIt2) scripts to "DetectHiddenWindows, off", so that things are more like AutoIt2. This is now the default for all scripts. From what I can tell, unlike AutoIt2, AutoIt3 always detects hidden windows. Since this can sometimes cause problems with window matching and distinctness, the AutoIt2 method seems to be a better default.

Changed the SetTitleMatchMode command. In addition to supporting the traditional modes of 1 & 2, the command now supports the words fast and slow (e.g. SetTitleMatchMode, slow). The default is fast, which is what AutoIt2 uses, whereas AutoIt3 uses the slow mode because it can "see" more text for certain windows. The fast mode performs significantly better, which may help the speed of scripts that that do a lot of windowing commands. The slow mode can be used to ensure full compatibility with AutoIt3 and its Window Spy (the Spy is included with AutoHotkey). I'm thinking of modifying the Window Spy so that it indicates which text of a window is available only in the "slow" mode.

Window finding reliability: Made the "slow" TitleMatchMode more reliable by increasing the timeout to 5 seconds because some windows are sluggish about checking their message queues (e.g. during long disk operations), and thus very slow to cough up their window text.

In this version, by default, multiple instances of the same script can be launched. If you wish to limit a script to a single instance, use the #SingleInstance directive described below.

ADDITIONS:

DriveSpaceFree (gets the free space [in MB] for a drive). For example:
DriveSpaceFree, OutputVar, C:\

SetAutoTrim, on/off (defaults to on): When you assign a value to a variable, this controls whether whitespace is automatically trimmed from the left and right of the string. For this purpose, whitespace consists of spaces and tabs, but NOT newlines (carriage returns). Thus, I believe the "on" setting yields the same behavior as AutoIt2 in this regard.

StringGetPos now supports a new optional last parameter. If the parameter is the letter R, the search goes from the right to left, so the last occurrence is found rather than the first.

ControlSetText: Derived from AutoIt3's function of the same name. For example:
ControlSetText, ControlName, Some new text, wintitle, wintext, exclude-title, exclude-text

Note that some or all of the last four parameters can be omitted. If all of them are omitted, the window found by the most recent IfWinExist, IfWinNotExist, or WinWait will be used. This behavior is the same for almost all of the other windowing commands too.

#AllowSameLineComments: To increase compatibility with AutoIt2, scripts that end in .aut are normally not permitted to have same-line comments (e.g. Run, notepad ; this is a comment). Add the line #AllowSameLineComments to the top of your script to allow them.

#SingleInstance: Specifying this anywhere in a script will prevent new instances of that script from being launched once there is already one running. Instead, you will be prompted for whether to close the existing instance. Note: You do not need to use "on" or "off" with this command.

#CommentFlag: Allows you to change the comment-flag from the default of semicolon to a string of your choice. The string can be up to 15 characters long. For example:
#CommentFlag // ; Change to C++ style.

#EscapeChar: The escape character is normally accent "`" but this allows you to change it to AutoIt2's escape character (backslash) or some other char of your choice. Note that by default, scripts that end in .aut will use the backslash for their escape char.

#DerefChar: The deref char is normally % (e.g. Sleep, %SleepTime%). You can change it to some other character with this.

TIPS

The first comma can always be omitted for any command, e.g. MsgBox This is the text of the Msgbox (single-parameter mode).

You can use the CapsLock keys as a new "modifier", similar to the windows key, without changing its native function (WinNT/2k/XP only, currently). For example:
CapsLock & A::run, notepad
CapsLock & S::run, calc