Version 0.211
Fixed the Reload command and the corresponding tray menu item to work with script filenames that contain spaces.

Added GetKeyState. For example:

GetKeyState, OutputVar, LButton

It sets OutputVar to be U (for up), D (for down), or blank if the state couldn't be determined for any reason. The key name can be virtually any of the keys listed in the documentation, including mouse buttons (at least under WinXP, and they will probably work on other OSes as well).

Added a new optional parameter to MouseClick. If this last parameter is D the button will be held down until it is released by the user physically clicking it or via another action in the script. If this last parameter is U, the button will be released (even if it wasn't down before, an up-event will still be sent).

Added new hotkey definition symbol ~ (tilde) to indicate that this hotkey's native function shouldn't be suppressed (hidden from the system) when the hotkey fires. For example, in the below, the mouse button click will be sent to the active window when it normally wouldn't be:

~RButton::Sleep, 1

Note: Special hotkeys that are substitutes for alt-tab always ignore this setting.

Changed commands that click the mouse (MouseClick, LeftClick, etc.) so that the clicks themselves cannot trigger hotkey subroutines. This prevents potential infinite loop conditions if a script were to send mouse events that happen to be valid hotkeys.

Changed slightly how hotkeys that are combinations of keyboard and mouse work: Normally if the prefix key of that pair modifies any other key (even something that isn't a hotkey) its key-up action won't fire (if it has one). This has been changed so that mouse buttons are the exception. For example:

Numpad0 & Numpad1::MsgBox, test1  ; Numpad0 is now a prefix key that is handled by the keyboard hook.
Numpad0::MsgBox, test2  ; Now Numpad0 is both a prefix AND a suffix (a suffix because it has its own action)

Under the old method, if Numpad0 was pressed down and then a mouse button clicked, the Numpad0 action would not fire when Numpad0 is released, as it normally would. Under the new method it does, but only if the combination of Numpad0 and that mouse button isn't a hotkey (since by default, the key/mouse events that comprise a hotkey are always hidden from the system unless the ~ symbol, described above, is used).