Version 0.215
Fixed a possible crash in variable assignments that involve more than one source. e.g. statements such as clipboard = %clipboard%%SomeVar%

Fixed a possible crash-on-startup bug introduced in version 0.212 in conjuction with "else if" support.

Fixed clipboard being hogged (held open) in rare cases such as displaying the clipboard with SplashTextOn.

Fixed PixelGetColor to accept negative coordinates without complaining.

Fixed a very obscure issue where a non-suppressed mouse hotkey such as "~LButton::" would cause that particular AutoHotkey instance's tray menu and main window context menus to behave strangely.

Added command SoundSetWaveVolume, which sets the playback volume for files such as WAV and MP3. For example:
SoundSetWaveVolume, 50 ; Number between 0 and 100, inclusive.

Added command EnvSet, which is needed because unlike AutoIt2, AutoHotkey does not store its variables in the environment. This is because performance would be worse and also because the OS limits env. vars to 32K (AutoHotkey vars, including the clipboard, are essentially unlimited in size). ErrorLevel is set to 1 if there was a problem or 0 otherwise. For example:
EnvSet, AutGUI, Some text to put in the variable.

Note: An EnvGet command is not necessary because dereferencing a blank or undefined variable will automatically trigger that behavior. For example, in the following line, the value of the ProgramFiles variable (if undefined in the script) will be fetched from the environment:
TargetFile = %ProgramFiles%\Something\setup.ini

Disabled the feature that jumps to the proper line in notepad or metapad when a syntax error is found (if the script is already open in such an editor) because it doesn't seem to work reliably. If any of you were using it, let me know. Later, I plan to add support for an external script such as "RunWhenSyntaxError.ahk" that will automatically be run if a script launch fails due to a syntax error.

Improved PixelSearch to set ErrorLevel to 2 if there was an error other than "color not found". It still sets it to 1 if the color isn't found, or 0 if it is.