Version 0.216
Fixed a bug introduced in a prior version that prevented new hotkeys from firing if the script had a MsgBox or other dialog displayed on the screen.

Fixed PixelSearch to accept negative coordinates without complaining.

Changed the behavior of scripts so that the default values of things such as KeyDelay and WinDelay will go into effect for every newly launched hotkey subroutine. You can use the auto-execute section -- which consists of the commands at the top of the script that occur prior to the first hotkey label -- to change any of the defaults that you wish. So now, when you press a hotkey, the subroutine will always start off fresh with the default values of the following:

DetectHiddenWindows
DetectHiddenText
SetTitleMatchMode
SetBatchLines
SetWinDelay
SetKeyDelay
SetDefaultMouseSpeed
SetStoreCapslockMode
AutoTrim
StringCaseSense

Added StringUpper and StringLower. Example: StringUpper, OutputVar, InputVar
Note: As with all such commands, InputVar and OutputVar are allowed to be the same variable.

Added the following built-in variables:

A_ThisHotkey: This is the label name (for example: #c) of the hotkey subroutine most recently launched. It's blank if none.
A_PriorHotkey: Same as above except for the subroutine that ran prior to this one.
A_TimeSinceThisHotkey: Milliseconds since the most recently launched hotkey subroutine began. It's -1 if
A_ThisHotkey is blank.
A_TimeSincePriorHotkey: Milliseconds since the prior hotkey subroutine was launched. It's -1 if A_PriorHotkey is blank.
A_TickCount: Milliseconds since the computer was last restarted. If this value ever exceeds 4294967295 (49.7 days), it will reset back to zero and begin counting from there.