Version 0.217
Fixed variable assignment (SetEnv and the new style, var = value) to properly handle escape sequences. Lines such as this should now work properly: result = %result% %color% `%`n

Fixed the shift key getting stuck down in rare cases when Numpad hotkeys invoke the Send command to send the shift key. The fix is basically a workaround for the fact that the OS, driver, or BIOS generates physical shift-key events to ensure that Numpad keys that have a dual-state (Numpad1-9 and Dot) never have their keystrokes occur while the shift key is down. Thanks once again to Robert Yaklin for his help on tracking this down.

Fixed Numpad hotkeys (e.g. NumpadEnd) that have a counterpart on the main keyboard (e.g. End) so that the counterpart key does not fire the hotkey.

Added commands IniRead/Write/Delete:
IniRead, OutputVar, Filespec, Section, Key, Default (value to return if key not found)
IniWrite, Value, Filespec, Section, Key
IniDelete, Filespec, Section, Key
Note: For compatibility reasons, IniRead does not support the last parameter if the script is of type .aut. Also, if this last parameter is omitted, OutputVar will be set to ERROR if there was a problem. This mirrors AutoIt2's behavior. Finally, none of these commands set ErrorLevel; this mirrors the behavior of AutoIt2 and 3.

Added command EnvUpdate, which according to the AutoIt3 docs: "Refreshes the OS environment. Similar effect as logging off and then on again." The command sets ErrorLevel to 0 to indicate success or 1 otherwise. Note: Due to the nature of this feature, there is a slight chance that keyboard and mouse events will lag (be delayed) while the command is running if the keyboard/mouse hooks are installed.

Added a built-in variable A_Space which contains a single space character. This avoids the need for a workaround to get a variable to contain a space. Example:

String = String with spaces
IfInString, String, %A_SPACE%
MsgBox, A space was found.
; But to assign a naked space to a variable of your own, be sure to turn off AutoTrim so that the assignment won't remove all leading and trailing spaces:
AutoTrim, off
MySpace = %A_SPACE%

Added the Edit command, which opens the current script for editing, or activates an existing Editor window if it finds one it thinks may have the file open based on its window title. This command is the same feature as selecting "Edit This Script" from the tray menu.

Added the Pause command and corresponding tray menu item. Unlike Suspend -- which disables hotkeys entirely -- pause will freeze the currently running subroutine (if none, pause will have no effect). The allowed values for the command's first parameter are ON, OFF, and TOGGLE (change to the opposite state). If the first parameter is omitted, it will default to TOGGLE just like the Suspend command. New hotkeys can still be launched while the script is paused, but when their subroutines finish, the underlying interrupted subroutine will still be paused. The "Pause, off" command will unpause only the most recently paused subroutine. Each subroutine can be paused independently of the others. The color of the tray icon changes to black whenever the script's current subroutine is in a paused state.

Improved the Suspend feature:

It now reverts all hotkeys back to their native functions rather than doing nothing when they are pressed.
The tray icon is changed to an S whenever hotkeys are suspended.
The first parameter can now be PERMIT or TOGGLE, in addition to ON or OFF. TOGGLE does the same thing as omitting the first parameter (i.e. toggles to the opposite state). PERMIT does not change the state of suspension. Instead, when used as the first line of any subroutine, it flags that subroutine as having permission to run even when hotkeys are suspended. Note: As before, any subroutine that has Suspend as its first command will be permitted to run while hotkeys are suspended (i.e. the PERMIT value is just a way to do it without changing the suspend state).
Note: Suspend will not disable the AlwaysOn/AlwaysOff feature for Scroll/Caps/Numlock if the script is using it.
Note: Although it isn't officially paused, the script is always halted while its tray menu is displayed. However, new hotkeys can be launched even then.

Improved the behavior of #SingleInstance scripts when the user attempts to run another instance: The user is now prompted for whether to keep the original instance or replace it with the new one.

Improved PixelSearch by making the output variables optional. In addition, an optional parameter has been added, which if present is a number between 0 and 255 (inclusive) to indicate the allowed number of shades of variation in either direction for the intensity of the red, green, and blue components of the color. This can be helpful if the color you're looking for isn't always exactly the same shade. Example:
PixelSearch, FoundX, FoundY, %X1%, %Y1%, %X2%, %Y2%, %Color%, 10 ; Allow 10 shades of variation.
Note: If you specify 255 shades of variation, all colors will match. The default is 0 shades.