Version 0.218
Added a script compiler (makes an EXE from a script) by popular demand. It was a lot easier than I thought, thanks to Jon's (AutoIt author's) open & modular coding style. It's available on the download page. Note: Presently, .aut scripts are not supported because there's no way for the compiled script to figure out whether to run in AutoIt2 mode (backslash as escape char) or AutoHotkey mode ([`] as escape char). So if necessary, first auto-convert your .aut file to .ahk by renaming it to "whatever.aut.ahk" and then running it with AutoHotkey.exe.

Added support for command line parameters becoming script variables. The command line format is:
AutoHotkey.exe [switches] [Script Filename] [Script Parameters]
CompiledScript.exe [switches] [Script Parameters]
Switches can be zero or more of the following:
/f or /force -- Launch unconditionally, skipping any warning dialogs.
/r or /restart -- Indicate that the script is being restarted (this is also used by the Reload command, internally)
Script Filename can be omitted if there are no Script Parameters. If omitted, it will run (or prompt you to create) AutoHotkey.ini in the current directory.
Script Parameters can be any strings you want to pass into the script. Any parameter containing spaces must be enclosed in double quotes. Like AutoIt2, the script sees them as the variables %1%, %2%, and so on.

Added command #NoTrayIcon. Put it anywhere in a script to disable the showing of a tray icon. If you use this for a script that has hotkeys, you might want to bind a hotkey to the ExitApp command. Otherwise, there will be no easy way to exit the program (without restarting the computer or killing the process). Example: #x::ExitApp