Version 0.219
Fixed the Suspend command so that when suspension is turned off, all hotkeys are re-enabled. Previously, hotkeys handled by the hook were sometimes left disabled.

Changed the Suspend command so that the keyboard and mouse hooks are never removed if the corresponding directives (#InstallKeybdHook and #InstallMouseHook) are present in the script.

Changed InputBox: For non-AutoIt2 (.aut) scripts, it now sets ErrorLevel to 1 and the output variable to the value entered if the user presses the Cancel button. For AutoIt2 scripts, the behavior is the same: ErrorLevel is not changed and the output variable is set to blank whenever the user presses Cancel, even if text was entered.

Changed FileDelete to set ErrorLevel to be the number of files that could not be deleted (if a wildcard pattern was given).

Changed the KeyHistory command so that the filename is specified in the second parameter. So now it works this way:

KeyHistory  ; same as before (display short log in window)
KeyHistory, on, c:\KeyHistory.txt  ; Turn on logging, optionally specifying the file to log to.
KeyHistory, off  ; Turn off logging.
KeyHistory, toggle [, filename]  ; Toggles to the opposite state (on or off).

Improved the performance of KeyHistory when it logs to a file by having it leave the file open rather than closing and reopening it for every key/mouse event. This should also reduce wear & tear on the drive. The only disadvantage is that viewing a log file that is open for logging might show contents that are out-of-date due to file buffering. To work around this, temporarily turn off logging before viewing the file.

Improved the performance of Run and RunWait by having them attempt CreateProcess() prior to ShellExecute(), rather than the other way around. I believe CreateProcess() is much faster.

Removed the commands FileToggleHidden and FileSetDateModified. See below for their replacements.

Added the following commands. Click a link for details:
ControlGetFocus
FileSelectFolder
FileGetAttrib
FileSetAttrib
FileGetTime
FileSetTime
FileGetSize
FileGetVersion