Version 0.2.21
Fixed a Loop bug introduced in the prior version: If a Loop was directly beneath an IF or ELSE (i.e. not enclosed in a block), the flow of execution would be wrong, usually resulting in parts of the script being erroneously executed twice. [thanks to Steve Drown]

Fixed the custom icon feature of the Ahk2Exe script compiler.

Variable names may now contain the following punctuation (previously, only underscore was allowed):
# _ @ $ ? [ ]

Documented the fact that MsgBox supports having a default button other than the first (256 = 2nd, 512 = 3rd).

Added the ability to compare, add to, and subtract from dates and times. This is done via an optional 3rd parameter for EnvSub and EnvAdd, which can be either Seconds, Minutes, Hours, or Days (or just the first letter of each of these). The date-time format is YYYYMMDDHH24MISS, which is used by all date-time commands. Example:

var1 = 20050126
var2 = 20040126
EnvSub, var1, %var2%, days
MsgBox, %var1%  ; The answer will be 366 since 2004 is a leap year.
var1 =  ; Make it blank so that the below will use the current time instead.
var1 += 31, days
MsgBox, %var1%  ; The answer will be the date 31 days from now.