Version 0.2.23
Fixed math operating on numbers that had trailing spaces: If a variable contained something like "123 " (i.e. space at the end) it would be considered non-numeric when math was performed on it, thus resulting in the variable being made blank. Sorry if that messed up anyone; the prior release is what broke it.