Version 0.2.24
Changed (again) math operations that involve impure numbers (e.g. 123abc) so that they conform to AutoIt2's behavior: Strings that start with a number are that number for the purpose of the calculation, e.g. 123abc is considered to be 123. Strings that do not start with a number are considered to be zero for the calculation, e.g. 123 + abc = 123. This change does not affect the behavior of IfEqual, IfGreater and IfLess: They will still compare strings alphabetically if either one is not purely numeric, which is slightly different from AutoIt2.

Added syntax to support determining whether a variable contains a pure number, floating point number, or time. Example:
if var is [not] number
if var is [not] float
if var is [not] time
{
...actions to take...
}