Version 0.2.25
I'm hoping this can be the final beta prior to initial release. All the test scripts run fine, including complex scripts that run for 72+ hours continuously without any memory leaks or other problems. The only significant fixes remaining are for ControlSend (uppercase vs. lowercase problems) and FileMove (wildcard support). If you get time over the next few days, please try to give this latest version a workout, especially the below changes and additions.

Fixed FileSelectFile: Sometimes the dialog would not be displayed, usually when doing a multi-select. [thanks to Steve Drown]

Fixed (hopefully) a hang-on-exit when the script is terminated while it is displaying a MsgBox or other dialog (rare even then).

Fixed #InstallKeybdHook and #InstallMouseHook so that they force the hooks to be installed even if there are no hotkeys that need them. This is useful to expand the scope of what the KeyHistory command shows.

Customized the AutoIt3 Window Spy to show which window text is available only when SetTitleMatchMode is in the "slow" mode. It also omits the mouse's desktop coordinates since they aren't used by AutoHotkey.

Changed the KeyLog command:

In an effort to prevent AutoHotkey from becoming branded a key logger or trojan by various security firms and their software -- after which it might be very difficult to "un-brand" it -- the ability to log to a file has been disabled. A version with that feature is available upon request, or you can compile it yourself with that section enabled.
Renamed the command to be KeyHistory to avoid even the appearance of supporting malicious use. Please change your scripts accordingly.
Improved the command so that it also displays the info that only the log file used to show: elapsed time between keystrokes and the target window's title.

Renamed command ControlLeftClick to ControlClick so that other mouse buttons are supported. A simple search and replace of ControlLeftClick with ControlClick in your scripts should do the trick because the command assumes the left button if none is specified.

Improved the SHIFT-NUMPAD issue: SHIFT shouldn't get stuck down as often when used with any of the dual-state numpad keys. This is a compensation for the keyboard's driver-generated shift key events, which ensure that SHIFT isn't logically down whenever a numpad key is pressed or released. [thanks to Robert Yaklin]

Added command SetMouseDelay to change the delay that occurs after every mouse movement or click. Default 10.

Added directives #MaxThreads and #MaxThreadsPerHotkey. The latter defaults to 1, which means that new instances of a hotkey subroutine cannot be launched if that subroutine is already running. This default behavior is what most people would want since it helps prevent accidental double-presses. Thus, there's probably no need to change any of your scripts.

Added command FileInstall, which includes extra files when compiling a script.

Added #Include directive: By default, files are included only once even if multiple inclusions are encountered. Use #IncludeAgain to include a file more than once.

Added menu bar to the main window for convenient access to features such as viewing variables, lines, hotkeys, and key history.