Version 1.0.01
Fixed EnvSub's date-time math so that if a newer time is subtracted from an older one, the result will be reported correctly as a negative number rather than a large positive number.

Fixed a rare problem with naked modifier keys used as hotkeys. For example, if RControl (by itself) was a hotkey, sometimes the control key would get stuck down after a Send operation that involved the control key.

Fixed syntax error reporting: in rare cases such as Else If statements, the line above where the syntax error actually occurred was reported.

Added the built-in variable %0%, which contains the number of command line parameters that were passed into the script. This also improves compatibility with AutoIt2. [thanks BoBo]

Added command BlockInput, which prevents the user from doing anything with the keyboard, mouse, and maybe other input devices as well. [thanks Jonathan Bennett & beardboy]