Version 1.0.04
Fixed Win-L (hotkey to lock the computer) so that the Windows key is not considered to be down when you log back in. [thanks Jack Horsfield].

Fixed the Reload feature: If a MsgBox or other dialog was displayed by the script at the time it was reloaded, the application would hang. [thanks Christian Ziemski]

Fixed auto-execute section, which is the top part of the script prior to the first hotkey label: If it had at least one normal (non-hotkey) label, in some cases the script's first hotkey label (if it had one) was executed as part of the auto-execute section. [thanks Robert Yaklin]

Improved SoundSetWaveVolume to accept +/- as a prefix. These prefixes adjust the volume up or down relative to its current level. In addition, the command now accepts an optional 2nd parameter to indicate which device to operate upon (for systems with more than one sound device).

Improved the following commands to greatly reduce the keyboard and mouse lag caused by a hook-script performing a long operation: FileDelete, FileCopy, FileMove, FileReadLine, FileSetAttrib, FileSetTime, Send, and ControlSend.

Improved StringLower/Upper to support converting a string to title case. For example, "GONE with the WIND" would become "Gone With The Wind".

Added remote registry support to registry-loops, RegRead, RegWrite, and RegDelete. Simply prepend the computer name to the root key, as in this example: \\workstation01:HKEY_LOCAL_MACHINE

Added built-in variable %A_TimeIdle%, which is the number of milliseconds since the system last received keyboard, mouse, or other input. This is useful for determining whether the user is away. This variable will be blank unless the operating system is Windows 2000, XP, or beyond.

Added command UrlDownloadToFile, which downloads a file from the Internet. [thanks Jonathan Bennett & Jason Payam Ahdoot]

Added command SoundGetWaveVolume, which retrieves the current WaveOut volume level.

Added commands SoundGet and SoundSet, which allow just about any setting of any sound device to be retrieved or changed. Examples include mute, microphone, master, synth, aux, and line-in.