Version 1.0.05
Fixed possible crash of the Exit and ExitApp commands.

Fixed Ctrl-Alt-Delete on Windows 2000/NT so that the Ctrl and Alt keys aren't thought to be down upon returning. [thanks Julian Doak]

Fixed GetKeyState so that the physical state of the mouse buttons is properly reported when the mouse hook is installed.

Fixed StringLower so that it accepts the optional Title-Case parameter as intended. [thanks beardboy]

Changed clipboard handling so that the program will keep trying to open the clipboard for 800 rather than 400 milliseconds before giving up and displaying a warning dialog. [thanks Jason Payam Ahdoot]

Changed: Variable names are now listed in alphabetical order in the main window. As a side-effect, performance of arrays is now better.

Changed FileAppend so that it optionally allows the file to be written in binary vs. text mode. As far as I know, the only consequence of this is that newline (`n) will be written as an isolated linefeed (LF) rather than a carriage return + linefeed (CR+LF).

Improved InputBox to allow a custom size and position to be specified. In addition, the window can now be resized by dragging its borders. [thanks David Nuttall & beardboy]

Added support for {SLEEP} to the Send command. Also added SLEEP as an allowed hotkey, though the Sleep button on some keyboards might not be compatible with this.

Added built-in variable %A_Index%, which is the number of the current loop iteration. The first iteration is 1, the second is 2, etc.

Added built-in variable %A_IdleTimePhysical%, which is similar to %A_IdleTime% except that it ignores artificial input (the kind generated by scripts and programs), and thus may be more useful for detecting whether the user is truly present. This variable will be the same as %A_IdleTime% unless the keyboard or mouse hook (or both) is installed.

Added command StringSplit, which divides a string into an array of substrings using the specified delimiters.

Added new type of loop: string-parse. It retrieves substrings (fields) from a string one by one (similar to StringSplit but uses less memory).

Added new type of loop: file-read. It reads the contents of a file (line by line) with higher performance than FileReadLine.