Version 1.0.06
Fixed the following issues specific to Windows 95/98/Me  [thanks Jay D. Novak]:

The Send command will now release the CTRL, ALT, SHIFT, and WIN keys when it should. Previously, a hotkey such as "#+a::Send, hello" would incorrectly send HELLO in uppercase.
SetNumlockState and "Send, {Numlock}" should now work more reliably.
GetKeyState should now work more reliably.
The delays caused by the following commands and settings are now much more accurate, at least on Windows 98SE: Sleep, SetKeyDelay, SetMouseDelay, SetWinDelay, SetControlDelay. If you use Win9x, you might have to increase the delay values in some of your scripts if you want the delays to be as long as they were in the prior version.

Fixed the tray icon so that it doesn't stay black when a new hotkey subroutine is launched while the current thread is paused. But it will go back to being black when the new subroutine finishes.

Fixed Gosub's and Goto's that jump to a label whose name is contained in a variable reference. In this example, the variable will now be resolved each time the line is run rather than only the first time: Gosub, %MyLabel%

Fixed the keyboard hook so that it is always auto-installed in response to #InstallKeybdHook and for scripts that contain only non-standard hotkeys such as "Capslock & e". [thanks bobhurt]

Changed the Installer to: 1) register the EXEs that come with the program so that you don't need to specify a full path when running them directly; 2) create its program group in All Users rather than the current user; 3) correctly remember the last installation directory (if you're installing on top of an older version).

Changed program startup for scripts that use the keyboard hook so that the current state of the CTRL/ALT/SHIFT/WIN keys is taken into account.

Changed the auto-execute section so that if it takes a long time to complete (or never completes), the default values for settings such as SetKeyDelay are put into effect after 100 milliseconds. Thus, it's usually best to make any desired changes to the default values at the top of the script. This is only important for scripts that contain hotkeys or timers.

Improved SetBatchLines to optionally accept a number of milliseconds rather than a number of lines. The millisecond method is recommended for scripts whenever speed and cooperation are important. For example, on most systems a setting of 10ms will prevent the script from using any more than 50% of an idle CPU's time. This allows scripts to run quickly while maintaining a high level of cooperation with CPU sensitive tasks such as games and video capture/playback.

Improved hotkey buffering during critical operations (such as Send), and when a hotkey's #MaxThreadsPerHotkey setting has been reached. In other words, in most cases the program will save up hotkey presses until they can be executed. As a side-effect, this makes holding down a hotkey (to auto-repeat it) more responsive. [thanks Jay D. Novak]

Improved the performance of hook scripts that have high or infinite BatchLines.

Added the AutoScriptWriter (macro recorder) to the main installer. I think there's a problem with the save button (which will hopefully be fixed soon) but it seems to work fine otherwise. [thanks Larry Keys]

Added command SetTimer, which is where most of the development time was spent over the past 10 days. Timers are useful because they run asynchronously, meaning that they will run at the specified frequency (interval) even when the script is waiting for a window, displaying a dialog, or busy with another task. Possible uses for timers include taking some action when the user becomes idle (as reflected by %A_TimeIdle%) or closing unwanted windows the moment they appear. There are many other potential uses.